refactor: prepare queue to be used by errors
The queuing logic isn't unique to transactions. It makes more sense to
put it directly in the lib folder. This also allows us to use it for
errors which we might want to do in the future.
