+++ lib/instrumentation/index.js
@@ -9,7 +9,7 @@ var hook = require('require-in-the-middle')
 var semver = require('semver')
 
 var logger = require('../logger')
-var Queue = require('./queue')
+var Queue = require('../queue')
 var request = require('../request')
 var Transaction = require('./transaction')
 
@@ -158,5 +158,5 @@ Instrumentation.prototype._recoverTransaction = function (trans) {
 }
 
 Instrumentation.prototype.flush = function (cb) {
-  this._queue._flush(cb)
+  this._queue.flush(cb)
 }
