+++ lib/queue.js
@@ -30,8 +30,14 @@ function Queue (opts, onFlush) {
 
 Queue.prototype.add = function (obj) {
   this._items.push(obj)
-  if (this._maxQueueSize !== -1 && this._items.length >= this._maxQueueSize) this._flush()
+  if (this._maxQueueSize !== -1 && this._items.length >= this._maxQueueSize) this.flush()
   else if (!this._timeout) this._queueFlush()
+}
+
+Queue.prototype.flush = function (cb) {
+  debug('flushing queue')
+  this._onFlush(this._items, cb || noop)
+  this._clear()
 }
 
 Queue.prototype._queueFlush = function () {
@@ -44,16 +50,10 @@ Queue.prototype._queueFlush = function () {
 
   debug('setting timer to flush queue: %dms', ms)
   this._timeout = setTimeout(function () {
-    self._flush()
+    self.flush()
   }, ms)
   this._timeout.unref()
   boot = false
-}
-
-Queue.prototype._flush = function (cb) {
-  debug('flushing queue')
-  this._onFlush(this._items, cb || noop)
-  this._clear()
 }
 
 Queue.prototype._clear = function () {
